# Conflicts:
#	src/backend/common/common-storage/storage-service/src/test/kotlin/com/tencent/bkrepo/common/storage/core/cache/CachedFileWriterTest.kt
#	src/backend/maven/biz-maven/src/main/kotlin/com/tencent/bkrepo/maven/artifact/repository/MavenLocalRepository.kt
#	src/backend/repository/api-repository/src/main/kotlin/com/tencent/bkrepo/repository/api/NodeClient.kt
#	src/backend/rpm/biz-rpm/build.gradle.kts

+++ src/backend/maven/biz-maven/src/main/kotlin/com/tencent/bkrepo/maven/service/MavenMetadataService.kt
@@ -95,16 +95,20 @@ class MavenMetadataService(
 
     fun delete(mavenArtifactInfo: ArtifactInfo, node: NodeDetail? = null, mavenGavc: MavenGAVC? = null) {
         node?.let {
-            val (criteria, _) = nodeCriteria(
+            val (criteria, mavenVersion) = nodeCriteria(
                 projectId = node.projectId,
                 repoName = node.repoName,
                 metadata = node.metadata,
                 fullPath = node.fullPath
             )
-            criteria?.let {
-                val query = Query(criteria)
-                mavenMetadataDao.remove(query)
+            if (criteria == null) return
+            criteria.apply {
+                mavenVersion?.timestamp?.let {
+                    this.and(TMavenMetadataRecord::timestamp.name).`is`(mavenVersion.timestamp)
+                }
             }
+            val query = Query(criteria)
+            mavenMetadataDao.remove(query)
         }
         mavenGavc?.let {
             val groupId = mavenGavc.groupId
